+++ scripts/inject-selector-script.py
@@ -39,8 +39,9 @@ def main(folder: str, selector_name: str) -> None:
 if __name__ == "__main__":
     logging.basicConfig(level=logging.INFO)
     try:
-        from fire import Fire
+        from jsonargparse import auto_cli, set_parsing_settings
 
-        Fire(main)
+        set_parsing_settings(parse_optionals_as_positionals=True)
+        auto_cli(main)
     except (ModuleNotFoundError, ImportError):
         main(*sys.argv[1:])